+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-# 0.2 (beta)
+# 0.2 (2020-08-10)
 
 - Added support for `$BROWSER` environment variable on CI.
 - Added `fake.person.honorific` property.
+- Changed shell command from `dev` to `shell`.
 
 # 0.1 (2020-07-03)